All commands are now separated, and are extensions of a superclass
