+++ src/com/dogonfire/gods/commands/CommandHome.java
@@ -1,0 +1,42 @@
+package com.dogonfire.gods.commands;
+
+import org.bukkit.ChatColor;
+import org.bukkit.Location;
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
+
+import com.dogonfire.gods.managers.BelieverManager;
+import com.dogonfire.gods.managers.GodManager;
+
+public class CommandHome extends GodsCommand {
+
+	protected CommandHome() {
+		super("home");
+		this.permission = "gods.home";
+	}
+
+	@Override
+	public void onCommand(CommandSender sender, String command, String... args) {
+		if (!hasPermission(sender)) {
+			sender.sendMessage(stringNoPermission);
+			return;
+		}
+		if (sender instanceof Player == false) {
+			sender.sendMessage(stringPlayerOnly);
+			return;
+		}
+		Player player = (Player) sender;
+		String godName = BelieverManager.get().getGodForBeliever(player.getUniqueId());
+
+		if (godName == null) {
+			sender.sendMessage(ChatColor.RED + "You do not believe in a God");
+			return;
+		}
+		Location location = GodManager.get().getHomeForGod(godName);
+		if (location == null) {
+			return;
+		}
+		player.teleport(location);
+	}
+
+}
